a few more sqlite specifics moved out of transactioner
