+++ chia/util/sqlite_wrapper.py
@@ -2,8 +2,12 @@ from __future__ import annotations
 
 import contextlib
 import functools
+import secrets
+import sqlite3
+import sys
 from collections.abc import AsyncIterator, Iterable
 from dataclasses import dataclass
+from datetime import datetime
 from pathlib import Path
 from types import TracebackType
 from typing import TYPE_CHECKING, Any, ClassVar, Optional, TextIO, TypeAlias, Union, cast
@@ -20,13 +24,64 @@ from chia.util.transactioner import (
     InternalError,
     Transactioner,
     manage_connection,
-    sql_trace_callback,
 )
 
 SqliteTransactioner: TypeAlias = Transactioner["SqliteConnection", "UntransactionedSqliteConnection"]
 
 # if TYPE_CHECKING:
 #     _protocol_check: AwaitableEnterable[Cursor] = cast(aiosqlite.Cursor, None)
+
+
+if aiosqlite.sqlite_version_info < (3, 32, 0):
+    SQLITE_MAX_VARIABLE_NUMBER = 900
+else:
+    SQLITE_MAX_VARIABLE_NUMBER = 32700
+
+# integers in sqlite are limited by int64
+SQLITE_INT_MAX = 2**63 - 1
+
+
+def generate_in_memory_db_uri() -> str:
+    # We need to use shared cache as our DB wrapper uses different types of connections
+    return f"file:db_{secrets.token_hex(16)}?mode=memory&cache=shared"
+
+
+# async def execute_fetchone(
+#     c: aiosqlite.Connection, sql: str, parameters: Optional[Iterable[Any]] = None
+# ) -> Optional[sqlite3.Row]:
+#     rows = await c.execute_fetchall(sql, parameters)
+#     for row in rows:
+#         return row
+#     return None
+
+
+def sql_trace_callback(req: str, file: TextIO, name: Optional[str] = None) -> None:
+    timestamp = datetime.now().strftime("%H:%M:%S.%f")
+    if name is not None:
+        line = f"{timestamp} {name} {req}\n"
+    else:
+        line = f"{timestamp} {req}\n"
+    file.write(line)
+
+
+def get_host_parameter_limit() -> int:
+    # NOTE: This does not account for dynamically adjusted limits since it makes a
+    #       separate db and connection.  If aiosqlite adds support we should use it.
+    if sys.version_info >= (3, 11):
+        connection = sqlite3.connect(":memory:")
+
+        limit_number = sqlite3.SQLITE_LIMIT_VARIABLE_NUMBER
+        host_parameter_limit = connection.getlimit(limit_number)
+    else:
+        # guessing based on defaults, seems you can't query
+
+        # https://www.sqlite.org/changes.html#version_3_32_0
+        # Increase the default upper bound on the number of parameters from 999 to 32766.
+        if sqlite3.sqlite_version_info >= (3, 32, 0):
+            host_parameter_limit = 32766
+        else:
+            host_parameter_limit = 999
+    return host_parameter_limit
 
 
 # TODO: think about the inheritance
@@ -83,6 +138,7 @@ class SqliteConnection:
         )
 
     _connection: aiosqlite.Connection
+    host_parameter_limit: ClassVar[int] = get_host_parameter_limit()
 
     async def close(self) -> None:
         return await self._connection.close()
